moved formatLang monkey patch in misc.py like in Odoo codebase
