+++ base_usability/models/__init__.py
@@ -3,4 +3,3 @@ from . import partner
 from . import company
 from . import mail
 from . import misc
-from . import report_sxw
